+++ noodles-bam/src/reader.rs
@@ -426,8 +426,12 @@ where
     fn alignment_records<'a>(
         &'a mut self,
         header: &'a sam::Header,
-    ) -> Box<dyn Iterator<Item = io::Result<RecordBuf>> + 'a> {
-        Box::new(self.records(header))
+    ) -> Box<dyn Iterator<Item = io::Result<Box<dyn sam::alignment::Record>>> + 'a> {
+        Box::new(
+            self.records(header).map(|result| {
+                result.map(|record| Box::new(record) as Box<dyn sam::alignment::Record>)
+            }),
+        )
     }
 }